@@ -1,31 +1,33 @@
 {
+  "name": "layer0-static-react-example",
   "scripts": {
-    "layer0:dev": "layer0 dev",
-    "layer0:build": "npm run build && layer0 build",
-    "layer0:deploy": "npm run build && layer0 deploy layer0-docs",
-    "layer0": "layer0",
-    "deploy": "npm run build && layer0 deploy",
-    "start": "react-scripts start",
+    "start": "BROWSER=none react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
-    "prettier:format": "prettier '**/*' --ignore-unknown --write"
+    "layer0:dev": "layer0 dev",
+    "layer0:build": "layer0 build",
+    "layer0:deploy": "layer0 deploy",
+    "layer0:production": "layer0 run --production",
+    "format": "prettier --write '**/*' --ignore-unknown"
   },
-  "name": "layer-stevie",
-  "version": "0.1.0",
-  "private": true,
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "^4.0.3"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1"
   },
   "devDependencies": {
-    "@layer0/cli": "^3.17.7",
-    "@layer0/core": "^3.17.7",
-    "@layer0/devtools": "^3.17.7",
-    "@layer0/prefetch": "^3.17.7",
-    "prettier": "^2.4.1"
+    "@heroicons/react": "^1.0.6",
+    "@layer0/cli": "^4.18.2",
+    "@layer0/core": "^4.18.2",
+    "@layer0/devtools": "^4.18.2",
+    "@layer0/prefetch": "^4.18.2",
+    "@layer0/react": "^4.18.2",
+    "@layer0/rum": "^3.1.0",
+    "autoprefixer": "^10.4.8",
+    "classnames": "^2.3.1",
+    "postcss": "^8.4.16",
+    "prettier": "^2.6.2",
+    "tailwindcss": "^3.1.8"
   },
   "browserslist": {
     "production": [
